Production ready website image (#8)
* Serve an optimized build of SnapShot

* Handle API responses better

* Separate dependencies from source in Dockerfile

* Build static files on container startup

This is counter-intuitive and definitely a trade-off worth explaining.

create-react-app keeps things simple by resolving environment variables
at build-time. Some variables are used to optimize the build itself,
so it's important that they are included for that stage. The caveat is
that we can't set environment variables after the build stage, which
makes runtime configuration impossible.

This quick compromise is to just build at runtime; which is admittedly
not lightweight. This way we can apply configuration at runtime though,
which gives us the flexibility we need. For high-performance situations,
it is probably reasonable to build derivative images with fixed
configurations.

* Run website as unprivileged user